refactor: rename to publicationHelpers
