+++ apps/web/src/components/Publication/Actions/Share/index.tsx
@@ -4,7 +4,7 @@ import { ArrowsRightLeftIcon } from '@heroicons/react/24/outline';
 import type { AnyPublication } from '@lenster/lens';
 import humanize from '@lenster/lib/humanize';
 import nFormatter from '@lenster/lib/nFormatter';
-import { isMirrorPublication } from '@lenster/lib/publicationTypes';
+import { isMirrorPublication } from '@lenster/lib/publicationHelpers';
 import stopEventPropagation from '@lenster/lib/stopEventPropagation';
 import { Spinner, Tooltip } from '@lenster/ui';
 import cn from '@lenster/ui/cn';
