+++ apps/web/src/components/Publication/PublicationStats.tsx
@@ -9,7 +9,7 @@ import {
 import { PUBLICATION } from '@lenster/data/tracking';
 import type { AnyPublication } from '@lenster/lens';
 import nFormatter from '@lenster/lib/nFormatter';
-import { isMirrorPublication } from '@lenster/lib/publicationTypes';
+import { isMirrorPublication } from '@lenster/lib/publicationHelpers';
 import { Modal } from '@lenster/ui';
 import { Leafwatch } from '@lib/leafwatch';
 import { Plural, t } from '@lingui/macro';
