+++ packages/lib/publicationHelpers.ts
